Avoid creating an EventManager
If we create the EventManager ourselves, it is getting in our way when we try to
re-use an existing DBAL connection: The DBAL Connection is holding the autoritative
EventManager instance. When calling EntityManager::create() and passing in a
pre-existing connection as well as the EventManager, this method will make sure both
EventManagers are the same.

So, we simply avoid creating the EventManager and instead provide the
one contained in our EntityManager.

This requires a little tweaking to how entity mappings can be registered
through the ResolveTargetEntityListener: We can create the ResolveTargetEntityListener
immediately when constructing the ORMInfrastructure, but have to postpone
registering it in the EventManager to the point where that is actually available.