Deprecate the `Query::getExecutionTimeInSeconds` method (#52)
The `\Doctrine\DBAL\Logging\DebugStack` class is used by the
`ORMInfrastructure` class to to collect the queries being executed.
`DebugStack` has been deprecated in DBAL 3.2
(https://github.com/doctrine/dbal/pull/4967) and we will need to replace
it with a logging middleware in the long run.

The logging middleware does not provide the query execution time, and it
would be burdensome to implement this metric. Thus, for the time being,
let's deprecate accessing the query execution time and remove this
feature in the next major.